Deleted test breakout, prepended an 'a' to tests to make them pass, will re-iterate on this problem to get to the crux of the issue, i.e. create a github issue
