+++ tests/test_integration.py
@@ -271,7 +271,7 @@ class TestViews(common.IntegrationTestCase):
         self.assertEqual(existing.total, total)
         self.assertEqual(existing.count, count)
 
-    def test_update_aggregate_size_create(self):
+    def test_update_aggregate_asize_create(self):
         filing = self.SchedAFactory(
             rpt_yr=2015,
             cmte_id='C6789',
@@ -298,7 +298,7 @@ class TestViews(common.IntegrationTestCase):
         self.assertEqual(rows[0].total, 0)
         self.assertEqual(rows[0].count, 0)
 
-    def test_update_aggregate_size_existing(self):
+    def test_update_aggregate_asize_existing(self):
         def get_existing():
             return models.ScheduleABySize.query.filter_by(
                 size=500,
@@ -359,3 +359,54 @@ class TestViews(common.IntegrationTestCase):
         # Updated total includes new Schedule A filing and new report
         self.assertAlmostEqual(existing.total, total + 75 + 20)
         self.assertEqual(existing.count, None)
+
+    def test_update_aggregate_purpose_create(self):
+        filing = self.SchedBFactory(
+            rpt_yr=2015,
+            cmte_id='C12345',
+            disb_amt=538,
+            disb_desc='CAMPAIGN BUTTONS',
+        )
+        db.session.commit()
+        manage.update_aggregates()
+        rows = models.ScheduleBByPurpose.query.filter_by(
+            cycle=2016,
+            committee_id='C12345',
+            purpose='MATERIALS',
+        ).all()
+        self.assertEqual(len(rows), 1)
+        self.assertEqual(rows[0].total, 538)
+        self.assertEqual(rows[0].count, 1)
+        filing.disbursement_description = 'BUMPER STICKERS'
+        db.session.add(filing)
+        db.session.commit()
+        manage.update_aggregates()
+        db.session.refresh(rows[0])
+        self.assertEqual(rows[0].total, 538)
+        self.assertEqual(rows[0].count, 1)
+        filing.disb_desc = 'HANGING OUT'
+        db.session.add(filing)
+        db.session.commit()
+        manage.update_aggregates()
+        db.session.refresh(rows[0])
+        self.assertEqual(rows[0].total, 0)
+        self.assertEqual(rows[0].count, 0)
+
+    def test_update_aggregate_purpose_existing(self):
+        existing = models.ScheduleBByPurpose.query.filter_by(
+            purpose='CONTRIBUTIONS',
+            cycle=2016,
+        ).first()
+        total = existing.total
+        count = existing.count
+        self.SchedBFactory(
+            rpt_yr=2015,
+            cmte_id=existing.committee_id,
+            disb_amt=538,
+            disb_tp='24K',
+        )
+        db.session.commit()
+        manage.update_aggregates()
+        db.session.refresh(existing)
+        self.assertEqual(existing.total, total + 538)
+        self.assertEqual(existing.count, count + 1)
